style: rename `error` to `server-error`
